requirements.txt
psycopg2==2.8.1
pandas==0.24.2
numpy==1.16.2
matplotlib==3.0.3
SQLAlchemy==1.3.2
ipykernel>=5.1.0
jupyterlab>=0.35.4
python-dotenv>=0.10.1
